refactor cm -> ui_cm_interface  for sciter and flutter
