+++ src/flutter_ffi.rs
@@ -13,7 +13,6 @@ use hbb_common::{
     fs, log,
 };
 
-use crate::flutter::connection_manager::{self, get_clients_length, get_clients_state};
 use crate::flutter::{self, SESSIONS};
 use crate::start_server;
 use crate::ui_interface;
@@ -673,13 +672,13 @@ pub fn main_get_online_statue() -> i64 {
     ONLINE.lock().unwrap().values().max().unwrap_or(&0).clone()
 }
 
-pub fn main_get_clients_state() -> String {
-    get_clients_state()
+pub fn cm_get_clients_state() -> String {
+    crate::ui_cm_interface::get_clients_state()
 }
 
-pub fn main_check_clients_length(length: usize) -> Option<String> {
-    if length != get_clients_length() {
-        Some(get_clients_state())
+pub fn cm_check_clients_length(length: usize) -> Option<String> {
+    if length != crate::ui_cm_interface::get_clients_length() {
+        Some(crate::ui_cm_interface::get_clients_state())
     } else {
         None
     }
@@ -791,27 +790,31 @@ pub fn main_get_mouse_time() -> f64 {
 }
 
 pub fn cm_send_chat(conn_id: i32, msg: String) {
-    connection_manager::send_chat(conn_id, msg);
+    crate::ui_cm_interface::send_chat(conn_id, msg);
 }
 
 pub fn cm_login_res(conn_id: i32, res: bool) {
-    connection_manager::on_login_res(conn_id, res);
+    if res {
+        crate::ui_cm_interface::authorize(conn_id);
+    } else {
+        crate::ui_cm_interface::close(conn_id);
+    }
 }
 
 pub fn cm_close_connection(conn_id: i32) {
-    connection_manager::close_conn(conn_id);
+    crate::ui_cm_interface::close(conn_id);
 }
 
 pub fn cm_check_click_time(conn_id: i32) {
-    connection_manager::check_click_time(conn_id)
+    crate::ui_cm_interface::check_click_time(conn_id)
 }
 
 pub fn cm_get_click_time() -> f64 {
-    connection_manager::get_click_time() as _
+    crate::ui_cm_interface::get_click_time() as _
 }
 
 pub fn cm_switch_permission(conn_id: i32, name: String, enabled: bool) {
-    connection_manager::switch_permission(conn_id, name, enabled)
+    crate::ui_cm_interface::switch_permission(conn_id, name, enabled)
 }
 
 pub fn main_get_icon() -> String {
