+++ src/ui.rs
@@ -125,7 +125,7 @@ pub fn start(args: &mut [String]) {
         page = "install.html";
     } else if args[0] == "--cm" {
         frame.register_behavior("connection-manager", move || {
-            Box::new(cm::ConnectionManager::new())
+            Box::new(cm::SciterConnectionManager::new())
         });
         page = "cm.html";
     } else if (args[0] == "--connect"
