+++ src/ui/cm.rs
@@ -1,60 +1,83 @@
 #[cfg(target_os = "linux")]
 use crate::ipc::start_pa;
-use crate::ipc::{self, new_listener, Connection, Data};
-use crate::VERSION;
+use crate::ui_cm_interface::{start_ipc, ConnectionManager, InvokeUiCM};
+
 #[cfg(windows)]
 use clipboard::{
     create_cliprdr_context, empty_clipboard, get_rx_clip_client, server_clip_file, set_conn_enabled,
 };
-use hbb_common::fs::{
-    can_enable_overwrite_detection, get_string, is_write_need_confirmation, new_send_confirm,
-    DigestCheckResult,
-};
-use hbb_common::{
-    allow_err,
-    config::Config,
-    fs, get_version_number, log,
-    message_proto::*,
-    protobuf::Message as _,
-    tokio::{self, sync::mpsc, task::spawn_blocking},
-};
-use sciter::{make_args, Element, Value, HELEMENT};
-use std::{
-    collections::HashMap,
-    ops::Deref,
-    sync::{Arc, RwLock},
-};
 
-pub struct ConnectionManagerInner {
-    root: Option<Element>,
-    senders: HashMap<i32, mpsc::UnboundedSender<Data>>,
-    click_time: i64,
+use hbb_common::{allow_err, log};
+use sciter::{make_args, Element, Value, HELEMENT};
+use std::sync::Mutex;
+use std::{ops::Deref, sync::Arc};
+
+#[derive(Clone, Default)]
+pub struct SciterHandler {
+    pub element: Arc<Mutex<Option<Element>>>,
 }
 
-#[derive(Clone)]
-pub struct ConnectionManager(Arc<RwLock<ConnectionManagerInner>>);
+impl InvokeUiCM for SciterHandler {
+    fn add_connection(&self, client: &crate::ui_cm_interface::Client) {
+        self.call(
+            "addConnection",
+            &make_args!(
+                client.id,
+                client.is_file_transfer,
+                client.port_forward.clone(),
+                client.peer_id.clone(),
+                client.name.clone(),
+                client.authorized,
+                client.keyboard,
+                client.clipboard,
+                client.audio,
+                client.file,
+                client.restart
+            ),
+        );
+    }
 
-impl Deref for ConnectionManager {
-    type Target = Arc<RwLock<ConnectionManagerInner>>;
+    fn remove_connection(&self, id: i32) {
+        self.call("removeConnection", &make_args!(id));
+        if crate::ui_cm_interface::get_clients_length().eq(&0) {
+            crate::platform::quit_gui();
+        }
+    }
+
+    fn new_message(&self, id: i32, text: String) {
+        self.call("newMessage", &make_args!(id, text));
+    }
+}
+
+impl SciterHandler {
+    #[inline]
+    fn call(&self, func: &str, args: &[Value]) {
+        if let Some(e) = self.element.lock().unwrap().as_ref() {
+            allow_err!(e.call_method(func, &super::value_crash_workaround(args)[..]));
+        }
+    }
+}
+
+pub struct SciterConnectionManager(ConnectionManager<SciterHandler>);
+
+impl Deref for SciterConnectionManager {
+    type Target = ConnectionManager<SciterHandler>;
 
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-impl ConnectionManager {
+impl SciterConnectionManager {
     pub fn new() -> Self {
         #[cfg(target_os = "linux")]
         std::thread::spawn(start_pa);
-        let inner = ConnectionManagerInner {
-            root: None,
-            senders: HashMap::new(),
-            click_time: Default::default(),
+        let cm = ConnectionManager {
+            ui_handler: SciterHandler::default(),
         };
-        let cm = Self(Arc::new(RwLock::new(inner)));
         let cloned = cm.clone();
         std::thread::spawn(move || start_ipc(cloned));
-        cm
+        SciterConnectionManager(cm)
     }
 
     fn get_icon(&mut self) -> String {
@@ -62,359 +85,27 @@ impl ConnectionManager {
     }
 
     fn check_click_time(&mut self, id: i32) {
-        let lock = self.read().unwrap();
-        if let Some(s) = lock.senders.get(&id) {
-            allow_err!(s.send(Data::ClickTime(0)));
-        }
+        crate::ui_cm_interface::check_click_time(id);
     }
 
     fn get_click_time(&self) -> f64 {
-        self.read().unwrap().click_time as _
-    }
-
-    #[inline]
-    fn call(&self, func: &str, args: &[Value]) {
-        let r = self.read().unwrap();
-        if let Some(ref e) = r.root {
-            allow_err!(e.call_method(func, &super::value_crash_workaround(args)[..]));
-        }
-    }
-
-    fn add_connection(
-        &self,
-        id: i32,
-        is_file_transfer: bool,
-        port_forward: String,
-        peer_id: String,
-        name: String,
-        authorized: bool,
-        keyboard: bool,
-        clipboard: bool,
-        audio: bool,
-        file: bool,
-        restart: bool,
-        tx: mpsc::UnboundedSender<Data>,
-    ) {
-        self.call(
-            "addConnection",
-            &make_args!(
-                id,
-                is_file_transfer,
-                port_forward,
-                peer_id,
-                name,
-                authorized,
-                keyboard,
-                clipboard,
-                audio,
-                file,
-                restart
-            ),
-        );
-        self.write().unwrap().senders.insert(id, tx);
-    }
-
-    fn remove_connection(&self, id: i32) {
-        self.write().unwrap().senders.remove(&id);
-        if self.read().unwrap().senders.len() == 0 {
-            crate::platform::quit_gui();
-        }
-        self.call("removeConnection", &make_args!(id));
-    }
-
-    async fn handle_data(
-        &self,
-        id: i32,
-        data: Data,
-        _tx_clip_file: &mpsc::UnboundedSender<ClipboardFileData>,
-        write_jobs: &mut Vec<fs::TransferJob>,
-        conn: &mut Connection,
-    ) {
-        match data {
-            Data::ChatMessage { text } => {
-                self.call("newMessage", &make_args!(id, text));
-            }
-            Data::ClickTime(ms) => {
-                self.write().unwrap().click_time = ms;
-            }
-            Data::FS(v) => match v {
-                ipc::FS::ReadDir {
-                    dir,
-                    include_hidden,
-                } => {
-                    Self::read_dir(&dir, include_hidden, conn).await;
-                }
-                ipc::FS::RemoveDir {
-                    path,
-                    id,
-                    recursive,
-                } => {
-                    Self::remove_dir(path, id, recursive, conn).await;
-                }
-                ipc::FS::RemoveFile { path, id, file_num } => {
-                    Self::remove_file(path, id, file_num, conn).await;
-                }
-                ipc::FS::CreateDir { path, id } => {
-                    Self::create_dir(path, id, conn).await;
-                }
-                ipc::FS::NewWrite {
-                    path,
-                    id,
-                    file_num,
-                    mut files,
-                    overwrite_detection,
-                } => {
-                    // cm has no show_hidden context
-                    // dummy remote, show_hidden, is_remote
-                    write_jobs.push(fs::TransferJob::new_write(
-                        id,
-                        "".to_string(),
-                        path,
-                        file_num,
-                        false,
-                        false,
-                        files
-                            .drain(..)
-                            .map(|f| FileEntry {
-                                name: f.0,
-                                modified_time: f.1,
-                                ..Default::default()
-                            })
-                            .collect(),
-                        overwrite_detection,
-                    ));
-                }
-                ipc::FS::CancelWrite { id } => {
-                    if let Some(job) = fs::get_job(id, write_jobs) {
-                        job.remove_download_file();
-                        fs::remove_job(id, write_jobs);
-                    }
-                }
-                ipc::FS::WriteDone { id, file_num } => {
-                    if let Some(job) = fs::get_job(id, write_jobs) {
-                        job.modify_time();
-                        Self::send(fs::new_done(id, file_num), conn).await;
-                        fs::remove_job(id, write_jobs);
-                    }
-                }
-                ipc::FS::CheckDigest {
-                    id,
-                    file_num,
-                    file_size,
-                    last_modified,
-                    is_upload,
-                } => {
-                    if let Some(job) = fs::get_job(id, write_jobs) {
-                        let mut req = FileTransferSendConfirmRequest {
-                            id,
-                            file_num,
-                            union: Some(file_transfer_send_confirm_request::Union::OffsetBlk(0)),
-                            ..Default::default()
-                        };
-                        let digest = FileTransferDigest {
-                            id,
-                            file_num,
-                            last_modified,
-                            file_size,
-                            ..Default::default()
-                        };
-                        if let Some(file) = job.files().get(file_num as usize) {
-                            let path = get_string(&job.join(&file.name));
-                            match is_write_need_confirmation(&path, &digest) {
-                                Ok(digest_result) => {
-                                    match digest_result {
-                                        DigestCheckResult::IsSame => {
-                                            req.set_skip(true);
-                                            let msg_out = new_send_confirm(req);
-                                            Self::send(msg_out, conn).await;
-                                        }
-                                        DigestCheckResult::NeedConfirm(mut digest) => {
-                                            // upload to server, but server has the same file, request
-                                            digest.is_upload = is_upload;
-                                            let mut msg_out = Message::new();
-                                            let mut fr = FileResponse::new();
-                                            fr.set_digest(digest);
-                                            msg_out.set_file_response(fr);
-                                            Self::send(msg_out, conn).await;
-                                        }
-                                        DigestCheckResult::NoSuchFile => {
-                                            let msg_out = new_send_confirm(req);
-                                            Self::send(msg_out, conn).await;
-                                        }
-                                    }
-                                }
-                                Err(err) => {
-                                    Self::send(fs::new_error(id, err, file_num), conn).await;
-                                }
-                            }
-                        }
-                    }
-                }
-                ipc::FS::WriteBlock {
-                    id,
-                    file_num,
-                    data,
-                    compressed,
-                } => {
-                    let raw = if let Ok(bytes) = conn.next_raw().await {
-                        Some(bytes)
-                    } else {
-                        None
-                    };
-                    if let Some(job) = fs::get_job(id, write_jobs) {
-                        if let Err(err) = job
-                            .write(
-                                FileTransferBlock {
-                                    id,
-                                    file_num,
-                                    data,
-                                    compressed,
-                                    ..Default::default()
-                                },
-                                raw.as_ref().map(|x| &x[..]),
-                            )
-                            .await
-                        {
-                            Self::send(fs::new_error(id, err, file_num), conn).await;
-                        }
-                    }
-                }
-                ipc::FS::WriteOffset {
-                    id: _,
-                    file_num: _,
-                    offset_blk: _,
-                } => {}
-            },
-            #[cfg(windows)]
-            Data::ClipbaordFile(_clip) => {
-                _tx_clip_file
-                    .send(ClipboardFileData::Clip((id, _clip)))
-                    .ok();
-            }
-            #[cfg(windows)]
-            Data::ClipboardFileEnabled(enabled) => {
-                _tx_clip_file
-                    .send(ClipboardFileData::Enable((id, enabled)))
-                    .ok();
-            }
-            _ => {}
-        }
-    }
-
-    async fn read_dir(dir: &str, include_hidden: bool, conn: &mut Connection) {
-        let path = {
-            if dir.is_empty() {
-                Config::get_home()
-            } else {
-                fs::get_path(dir)
-            }
-        };
-        if let Ok(Ok(fd)) = spawn_blocking(move || fs::read_dir(&path, include_hidden)).await {
-            let mut msg_out = Message::new();
-            let mut file_response = FileResponse::new();
-            file_response.set_dir(fd);
-            msg_out.set_file_response(file_response);
-            Self::send(msg_out, conn).await;
-        }
-    }
-
-    async fn handle_result<F: std::fmt::Display, S: std::fmt::Display>(
-        res: std::result::Result<std::result::Result<(), F>, S>,
-        id: i32,
-        file_num: i32,
-        conn: &mut Connection,
-    ) {
-        match res {
-            Err(err) => {
-                Self::send(fs::new_error(id, err, file_num), conn).await;
-            }
-            Ok(Err(err)) => {
-                Self::send(fs::new_error(id, err, file_num), conn).await;
-            }
-            Ok(Ok(())) => {
-                Self::send(fs::new_done(id, file_num), conn).await;
-            }
-        }
-    }
-
-    async fn remove_file(path: String, id: i32, file_num: i32, conn: &mut Connection) {
-        Self::handle_result(
-            spawn_blocking(move || fs::remove_file(&path)).await,
-            id,
-            file_num,
-            conn,
-        )
-        .await;
-    }
-
-    async fn create_dir(path: String, id: i32, conn: &mut Connection) {
-        Self::handle_result(
-            spawn_blocking(move || fs::create_dir(&path)).await,
-            id,
-            0,
-            conn,
-        )
-        .await;
-    }
-
-    async fn remove_dir(path: String, id: i32, recursive: bool, conn: &mut Connection) {
-        let path = fs::get_path(&path);
-        Self::handle_result(
-            spawn_blocking(move || {
-                if recursive {
-                    fs::remove_all_empty_dir(&path)
-                } else {
-                    std::fs::remove_dir(&path).map_err(|err| err.into())
-                }
-            })
-            .await,
-            id,
-            0,
-            conn,
-        )
-        .await;
-    }
-
-    async fn send(msg: Message, conn: &mut Connection) {
-        match msg.write_to_bytes() {
-            Ok(bytes) => allow_err!(conn.send(&Data::RawMessage(bytes)).await),
-            err => allow_err!(err),
-        }
+        crate::ui_cm_interface::get_click_time() as _
     }
 
     fn switch_permission(&self, id: i32, name: String, enabled: bool) {
-        let lock = self.read().unwrap();
-        if let Some(s) = lock.senders.get(&id) {
-            allow_err!(s.send(Data::SwitchPermission { name, enabled }));
-        }
+        crate::ui_cm_interface::switch_permission(id, name, enabled);
     }
 
     fn close(&self, id: i32) {
-        let lock = self.read().unwrap();
-        if let Some(s) = lock.senders.get(&id) {
-            allow_err!(s.send(Data::Close));
-        }
-    }
-
-    fn send_msg(&self, id: i32, text: String) {
-        let lock = self.read().unwrap();
-        if let Some(s) = lock.senders.get(&id) {
-            allow_err!(s.send(Data::ChatMessage { text }));
-        }
-    }
-
-    fn send_data(&self, id: i32, data: Data) {
-        let lock = self.read().unwrap();
-        if let Some(s) = lock.senders.get(&id) {
-            allow_err!(s.send(data));
-        }
+        crate::ui_cm_interface::close(id);
     }
 
     fn authorize(&self, id: i32) {
-        let lock = self.read().unwrap();
-        if let Some(s) = lock.senders.get(&id) {
-            allow_err!(s.send(Data::Authorize));
-        }
+        crate::ui_cm_interface::authorize(id);
+    }
+
+    fn send_msg(&self, id: i32, text: String) {
+        crate::ui_cm_interface::send_chat(id, text);
     }
 
     fn t(&self, name: String) -> String {
@@ -422,9 +113,9 @@ impl ConnectionManager {
     }
 }
 
-impl sciter::EventHandler for ConnectionManager {
+impl sciter::EventHandler for SciterConnectionManager {
     fn attached(&mut self, root: HELEMENT) {
-        self.write().unwrap().root = Some(Element::from(root));
+        *self.ui_handler.element.lock().unwrap() = Some(Element::from(root));
     }
 
     sciter::dispatch_script_call! {
@@ -436,181 +127,5 @@ impl sciter::EventHandler for ConnectionManager {
         fn authorize(i32);
         fn switch_permission(i32, String, bool);
         fn send_msg(i32, String);
-    }
-}
-
-pub enum ClipboardFileData {
-    #[cfg(windows)]
-    Clip((i32, ipc::ClipbaordFile)),
-    Enable((i32, bool)),
-}
-
-#[tokio::main(flavor = "current_thread")]
-async fn start_ipc(cm: ConnectionManager) {
-    let (tx_file, _rx_file) = mpsc::unbounded_channel::<ClipboardFileData>();
-    #[cfg(windows)]
-    let cm_clip = cm.clone();
-    #[cfg(windows)]
-    std::thread::spawn(move || start_clipboard_file(cm_clip, _rx_file));
-
-    #[cfg(windows)]
-    std::thread::spawn(move || {
-        log::info!("try create privacy mode window");
-        #[cfg(windows)]
-        {
-            if let Err(e) = crate::platform::windows::check_update_broker_process() {
-                log::warn!(
-                    "Failed to check update broker process. Privacy mode may not work properly. {}",
-                    e
-                );
-            }
-        }
-        allow_err!(crate::ui::win_privacy::start());
-    });
-
-    match new_listener("_cm").await {
-        Ok(mut incoming) => {
-            while let Some(result) = incoming.next().await {
-                match result {
-                    Ok(stream) => {
-                        log::debug!("Got new connection");
-                        let mut stream = Connection::new(stream);
-                        let cm = cm.clone();
-                        let tx_file = tx_file.clone();
-                        tokio::spawn(async move {
-                            // for tmp use, without real conn id
-                            let conn_id_tmp = -1;
-                            let mut conn_id: i32 = 0;
-                            let (tx, mut rx) = mpsc::unbounded_channel::<Data>();
-                            let mut write_jobs: Vec<fs::TransferJob> = Vec::new();
-                            loop {
-                                tokio::select! {
-                                    res = stream.next() => {
-                                        match res {
-                                            Err(err) => {
-                                                log::info!("cm ipc connection closed: {}", err);
-                                                break;
-                                            }
-                                            Ok(Some(data)) => {
-                                                match data {
-                                                    Data::Login{id, is_file_transfer, port_forward, peer_id, name, authorized, keyboard, clipboard, audio, file, file_transfer_enabled, restart} => {
-                                                        log::debug!("conn_id: {}", id);
-                                                        conn_id = id;
-                                                        tx_file.send(ClipboardFileData::Enable((id, file_transfer_enabled))).ok();
-                                                        cm.add_connection(id, is_file_transfer, port_forward, peer_id, name, authorized, keyboard, clipboard, audio, file, restart, tx.clone());
-                                                    }
-                                                    Data::Close => {
-                                                        tx_file.send(ClipboardFileData::Enable((conn_id, false))).ok();
-                                                        log::info!("cm ipc connection closed from connection request");
-                                                        break;
-                                                    }
-                                                    Data::PrivacyModeState((id, _)) => {
-                                                        conn_id = conn_id_tmp;
-                                                        cm.send_data(id, data)
-                                                    }
-                                                    _ => {
-                                                        cm.handle_data(conn_id, data, &tx_file, &mut write_jobs, &mut stream).await;
-                                                    }
-                                                }
-                                            }
-                                            _ => {}
-                                        }
-                                    }
-                                    Some(data) = rx.recv() => {
-                                        if stream.send(&data).await.is_err() {
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                            if conn_id != conn_id_tmp {
-                                cm.remove_connection(conn_id);
-                            }
-                        });
-                    }
-                    Err(err) => {
-                        log::error!("Couldn't get cm client: {:?}", err);
-                    }
-                }
-            }
-        }
-        Err(err) => {
-            log::error!("Failed to start cm ipc server: {}", err);
-        }
-    }
-    crate::platform::quit_gui();
-}
-
-#[cfg(windows)]
-#[tokio::main(flavor = "current_thread")]
-pub async fn start_clipboard_file(
-    cm: ConnectionManager,
-    mut rx: mpsc::UnboundedReceiver<ClipboardFileData>,
-) {
-    let mut cliprdr_context = None;
-    let mut rx_clip_client = get_rx_clip_client().lock().await;
-
-    loop {
-        tokio::select! {
-            clip_file = rx_clip_client.recv() => match clip_file {
-                Some((conn_id, clip)) => {
-                    cmd_inner_send(
-                        &cm,
-                        conn_id,
-                        Data::ClipbaordFile(clip)
-                    );
-                }
-                None => {
-                    //
-                }
-            },
-            server_msg = rx.recv() => match server_msg {
-                Some(ClipboardFileData::Clip((conn_id, clip))) => {
-                    if let Some(ctx) = cliprdr_context.as_mut() {
-                        server_clip_file(ctx, conn_id, clip);
-                    }
-                }
-                Some(ClipboardFileData::Enable((id, enabled))) => {
-                    if enabled && cliprdr_context.is_none() {
-                        cliprdr_context = Some(match create_cliprdr_context(true, false) {
-                            Ok(context) => {
-                                log::info!("clipboard context for file transfer created.");
-                                context
-                            }
-                            Err(err) => {
-                                log::error!(
-                                    "Create clipboard context for file transfer: {}",
-                                    err.to_string()
-                                );
-                                return;
-                            }
-                        });
-                    }
-                    set_conn_enabled(id, enabled);
-                    if !enabled {
-                        if let Some(ctx) = cliprdr_context.as_mut() {
-                            empty_clipboard(ctx, id);
-                        }
-                    }
-                }
-                None => {
-                    break
-                }
-            }
-        }
-    }
-}
-
-#[cfg(windows)]
-fn cmd_inner_send(cm: &ConnectionManager, id: i32, data: Data) {
-    let lock = cm.read().unwrap();
-    if id != 0 {
-        if let Some(s) = lock.senders.get(&id) {
-            allow_err!(s.send(data));
-        }
-    } else {
-        for s in lock.senders.values() {
-            allow_err!(s.send(data.clone()));
-        }
     }
 }
